Fix Issue #564
astropy<1.1.0 uses a different value for 'u' when converting quantities
than the one specified as constants.u.
In astropy>=1.1 unit.u is changed to be equal to constants.u.
By using constants.u, our results will be independant of the astropy
version.

This change required an update of the reference-files for
test_plasmas_full.py and test_tardis_full.py.

Resolves: #564